backend/pyproject.toml
[project]
name = "kairoai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pipecat-ai[cartesia,deepgram,openai]>=1.1.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.2",
    "yt-dlp>=2026.3.17",
]
